Resuelto problema con las timezones que se dividian de dia, en los acompañantes
